requirements.txt
streamlit
pillow
numpy
tensorflow-cpu